requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
ansible-base==2.10.8
apache-libcloud==3.2.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
bottle==0.12.21
Brlapi==0.8.3
cattrs==22.1.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
cryptography==36.0.2
cupshelpers==1.0
cycler==0.11.0
dateparser==1.1.1
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.1.0
duplicity==0.8.21
entrypoints==0.4
erdpy==2.0.1
exceptiongroup==1.0.0rc8
executing==1.0.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.6.0
fonttools==4.37.1
frozenlist==1.3.1
future==0.18.2
hidapi==0.12.0.post2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.9.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ledgercomm==1.1.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.12
ntlm-auth==1.4.0
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.4
pandas-ta==0.3.14b0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.15.0
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-binance==1.0.16
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.3.2
reportlab==3.6.8
requests==2.25.1
requests-cache==0.9.5
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selinux==3.3
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.0
systemd-python==234
ta==0.10.2
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
traitlets==5.3.0
tzdata==2022.2
tzlocal==4.2
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ujson==5.4.0
unattended-upgrades==0.1
url-normalize==1.4.3
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
widgetsnbextension==4.0.3
xdg==5
xkit==0.0.0
xmltodict==0.12.0
yarl==1.8.1
zipp==1.0.0
